Split out CastableList and related types, and placed in core.
